% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{calc_gof}
\alias{calc_gof}
\title{Calculate a goodness-of-fit table}
\usage{
calc_gof(sim, obs, funs, period = NULL)
}
\arguments{
\item{sim}{Data frame with one date column and one or many columns with
values of the simulated variable.}

\item{obs}{Data frame with one date and one value column.}

\item{funs}{A list of functions used to calculate goodness-of-fit (GoF) values.
This can be a named or unnamed list. If named (e.g., \code{list(nse_q = NSE, pb_q = pbias)}),
the names will be used as column names in the returned results. If unnamed,
the function names themselves will be used as column names. Commonly used
functions (such as \code{NSE}, \code{pbias}, \code{KGE}, etc.) are available in the
documentation of the \emph{hydroGOF} package.}

\item{period}{Optional. A character string specifying the time period for
aggregating data before calculating GoF. One of \code{"second"}, \code{"minute"},
\code{"hour"}, \code{"day"}, \code{"week"}, \code{"month"}, \code{"bimonth"}, \code{"quarter"},
\code{"season"}, \code{"halfyear"}, \code{"year"}, or \code{NULL} (default, no aggregation).
Aggregation is done using the mean. Period names match the unit strings
accepted by \code{\link[lubridate:round_date]{lubridate::floor_date()}}.}
}
\value{
A table with the calculated goodness-of-fit values.
}
\description{
\code{calc_gof()} calculates goodness-of-fit inices for simulated time series
\code{sim} and observations \code{obs} applying the functions passed as a list
with \code{funs}.
}
\examples{
\dontrun{
date_flow <- seq(as.Date('2000-01-01'), as.Date('2000-12-31'), by = 'day')

flow_sim <- data.frame(date = date_flow,
                       run_1 = runif(length(date_flow)),
                       run_2 = runif(length(date_flow))
                       )
flow_obs <- data.frame(date = date_flow,
                       qobs = runif(length(date_flow))
                       )

obj_tbl <- calc_gof(sim = flow_sim, obs = flow_obs, funs = list(cor_q = cor))
}

}
\seealso{
For available goodness-of-fit functions, see the \href{https://cran.r-project.org/web/packages/hydroGOF/hydroGOF.pdf}{hydroGOF package documentation}.
}
\keyword{calculate}
